Shortens nocompile logic
