+++ pkg/flare/manifests.go
@@ -16,6 +16,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"os"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -55,7 +56,7 @@ func convertToYAMLBytes(input any) ([]byte, error) {
 
 // Retrieve a DaemonSet YAML from the API server for a given name and namespace, and returns the associated YAML manifest into a a byte array.
 // Its purpose is to retrieve the Datadog Agent DaemonSet manifest when building a Cluster Agent flare.
-func GetDaemonset(cl *apiserver.APIClient, name string, namespace string) ([]byte, error) {
+func getDaemonset(cl *apiserver.APIClient, name string, namespace string) ([]byte, error) {
 	ds, err := cl.Cl.AppsV1().DaemonSets(namespace).Get(context.TODO(), name, metav1.GetOptions{})
 	if err != nil {
 		log.Debugf("Can't retrieve DaemonSet %v from the API server: %s", name, err.Error())
@@ -64,15 +65,61 @@ func GetDaemonset(cl *apiserver.APIClient, name string, namespace string) ([]byt
 	return convertToYAMLBytes(ds)
 }
 
+// getAgentDaemonSet retrieves the DaemonSet manifest of the Agent
+func getAgentDaemonSet() ([]byte, error) {
+	// The Agent DaemonSet name is based on the Helm chart template and added to the Cluster Agent as an environment variable
+	var agentDaemonsetName string
+	var releaseNamespace string
+	var agentDaemonset []byte
+
+	cl, err := apiserver.GetAPIClient()
+	if err != nil {
+		return nil, log.Errorf("Can't create client to query the API Server: %s", err)
+	}
+	agentDaemonsetName = os.Getenv(HELM_AGENT_DAEMONSET)
+	releaseNamespace = os.Getenv(HELM_CHART_RELEASE_NAMESPACE)
+	if agentDaemonsetName == "" || releaseNamespace == "" {
+		return nil, log.Errorf("Can't collect the Agent Daemonset name and/or namespace from the environment variables %s and %v", HELM_AGENT_DAEMONSET, HELM_CHART_RELEASE_NAMESPACE)
+	}
+	agentDaemonset, err = getDaemonset(cl, agentDaemonsetName, releaseNamespace)
+	if err != nil {
+		return nil, log.Errorf("Error while collecting the Agent DaemonSet: %q", err)
+	}
+	return agentDaemonset, nil
+}
+
 // Retrieve a Deployment YAML from the API server for a given name and namespace, and returns the associated YAML manifest into a a byte array.
 // Its purpose is to retrieve the Datadog Cluster Agent Deployment manifest when building a Cluster Agent flare.
-func GetDeployment(cl *apiserver.APIClient, name string, namespace string) ([]byte, error) {
+func getDeployment(cl *apiserver.APIClient, name string, namespace string) ([]byte, error) {
 	deploy, err := cl.Cl.AppsV1().Deployments(namespace).Get(context.TODO(), name, metav1.GetOptions{})
 	if err != nil {
 		log.Debugf("Can't retrieve Deployment %v from the API server: %s", name, err.Error())
 		return nil, err
 	}
 	return convertToYAMLBytes(deploy)
+}
+
+// getClusterAgentDeployment retrieves the Deployment manifest of the Cluster Agent
+func getClusterAgentDeployment() ([]byte, error) {
+	// The Cluster Agent Deployment name is based on the Helm chart template and added to the Cluster Agent as an environment variable
+	var clusterAgentDeploymentName string
+	var releaseNamespace string
+	var clusterAgentDeployment []byte
+
+	cl, err := apiserver.GetAPIClient()
+	if err != nil {
+		return nil, log.Errorf("Can't create client to query the API Server: %s", err)
+	}
+	clusterAgentDeploymentName = os.Getenv(HELM_CLUSTER_AGENT_DEPLOYMENT)
+	releaseNamespace = os.Getenv(HELM_CHART_RELEASE_NAMESPACE)
+	if clusterAgentDeploymentName == "" || releaseNamespace == "" {
+		return nil, log.Errorf("Can't collect the Cluster Agent Deployment name and/or namespace from the environment variables %s and %v", HELM_CLUSTER_AGENT_DEPLOYMENT, HELM_CHART_RELEASE_NAMESPACE)
+	}
+	clusterAgentDeployment, err = getDeployment(cl, clusterAgentDeploymentName, releaseNamespace)
+	if err != nil {
+		return nil, log.Errorf("Error while collecting the Cluster Agent Deployment: %q", err)
+	}
+	return clusterAgentDeployment, nil
 }
 
 // getDeployedHelmConfigmap returns the configmap for a given release.
@@ -165,4 +212,50 @@ func decodeChartValuesFromRelease(encodedRelease string) ([]byte, error) {
 		return nil, err
 	}
 	return yaml.JSONToYAML(configjson)
+}
+
+// getHelmValues retrieves the user-defined values for the Datadog Helm chart
+func getHelmValues() ([]byte, error) {
+	var dataString string
+	var helmUserValues []byte
+	var releaseName string
+	var releaseNamespace string
+
+	cl, err := apiserver.GetAPIClient()
+	if err != nil {
+		return nil, log.Errorf("Can't create client to query the API Server: %s", err)
+	}
+	releaseName = os.Getenv(HELM_CHART_RELEASE_NAME)
+	releaseNamespace = os.Getenv(HELM_CHART_RELEASE_NAMESPACE)
+	if releaseName == "" || releaseNamespace == "" {
+		return nil, log.Errorf("Can't collect the Datadog Helm chart release name and/or namespace from the environment variables %s and %v", HELM_CHART_RELEASE_NAME, HELM_CHART_RELEASE_NAMESPACE)
+	}
+	// Attempting to retrieve Helm chart data from secrets (default storage in Helm v3)
+	secret, err := getDeployedHelmSecret(cl, releaseName, releaseNamespace)
+	if err != nil {
+		log.Warnf("Error while collecting the Helm chart values from secret: %v", err)
+	} else {
+		// Contrary to the Configmap, the secret data is a byte array, so the string function is necessary
+		dataString = string(secret.Data["release"])
+		helmUserValues, err = decodeChartValuesFromRelease(dataString)
+		if err != nil {
+			log.Warnf("Unable to decode release stored in secret: %v", err)
+		} else {
+			return helmUserValues, nil
+		}
+	}
+	// The cluster Agent was unable to retrieve Helm chart data from secrets, attempting to retrieve them from Configmaps
+	configmap, err := getDeployedHelmConfigmap(cl, releaseName, releaseNamespace)
+	if err != nil {
+		log.Warnf("Error while collecting the Helm chart values from configmap: %v", err)
+	} else {
+		dataString = configmap.Data["release"]
+		helmUserValues, err = decodeChartValuesFromRelease(dataString)
+		if err != nil {
+			log.Warnf("Unable to decode release stored in configmap: %v", err)
+		} else {
+			return helmUserValues, nil
+		}
+	}
+	return nil, fmt.Errorf("Unable to collect Helm values from secrets/configmaps")
 }
